backend/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
sqlalchemy
pydantic
